Always output to an intermediate file
I previously created an option to output to an intermediate file, but
I am not changing this to be the only behavior, as it is necessary for
large repositories.

Large content cannot be stored in RAM as was being done before. It now
always writes to a git-p4-dump file, which it then automatically feeds
into the fast-import. This allows repositories of any size to be
imported.